requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.29.0

# ORM
sqlalchemy==2.0.29
alembic==1.13.1

# Pydantic (compatibile con FastAPI)
pydantic==2.7.1
pydantic-settings==2.2.1

# Password hashing
passlib[bcrypt]==1.7.4

# JWT
python-jose==3.3.0

# Upload form-data
python-multipart==0.0.9

# Env
python-dotenv==1.0.1
